API_day1/API_1day/frontend/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.5.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
altair=5.5.0=pyhd8ed1ab_1
arrow-cpp=19.0.0=h865e1df_2
attrs=25.3.0=pyh71513ae_0
aws-c-auth=0.6.19=h5eee18b_0
aws-c-cal=0.5.20=hdbd6064_0
aws-c-common=0.8.5=h5eee18b_0
aws-c-compression=0.2.16=h5eee18b_0
aws-c-event-stream=0.2.15=h6a678d5_0
aws-c-http=0.6.25=h5eee18b_0
aws-c-io=0.13.10=h5eee18b_0
aws-c-mqtt=0.7.13=h5eee18b_0
aws-c-s3=0.1.51=hdbd6064_0
aws-c-sdkutils=0.1.6=h5eee18b_0
aws-checksums=0.1.13=h5eee18b_0
aws-crt-cpp=0.18.16=h6a678d5_0
aws-sdk-cpp=1.11.212=hecad206_0
blas=2.131=openblas
blas-devel=3.9.0=31_h1ea3ea9_openblas
blinker=1.9.0=pyhff2d567_0
boost-cpp=1.82.0=hdb19cb5_2
bottleneck=1.4.2=py39ha9d4c09_0
brotli=1.0.9=h166bdaf_7
brotli-bin=1.0.9=h166bdaf_7
brotli-python=1.0.9=py39h5a03fae_7
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2025.6.15=hbd8a1cb_0
cachetools=6.1.0=pyhd8ed1ab_0
certifi=2025.6.15=pyhd8ed1ab_0
cffi=1.15.0=py39h4bc2ebd_0
charset-normalizer=3.4.2=pyhd8ed1ab_0
click=8.1.8=pyh707e725_0
contourpy=1.2.1=py39hdb19cb5_1
cycler=0.12.1=pyhd8ed1ab_1
cyrus-sasl=2.1.28=h52b45da_1
dbus=1.13.18=hb2f20db_0
expat=2.7.1=h6a678d5_0
filelock=3.13.1=pypi_0
fontconfig=2.14.1=h55d465d_3
fonttools=4.33.3=py39hb9d737c_0
freetype=2.13.3=h4a9f257_0
fsspec=2025.5.1=pypi_0
gflags=2.2.2=he1b5a44_1004
gitdb=4.0.12=pyhd8ed1ab_0
gitpython=3.1.44=pyhff2d567_0
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
glog=0.5.0=h48cff8f_0
h2=4.2.0=pyhd8ed1ab_0
hf-xet=1.1.5=pypi_0
hpack=4.1.0=pyhd8ed1ab_0
huggingface-hub=0.33.0=pypi_0
hyperframe=6.1.0=pyhd8ed1ab_0
icu=73.1=h6a678d5_0
idna=3.10=pyhd8ed1ab_1
importlib-metadata=8.7.0=pyhe01879c_1
importlib_resources=6.5.2=pyhd8ed1ab_0
jinja2=3.1.6=pyhd8ed1ab_0
jpeg=9e=h166bdaf_1
jsonschema=4.24.0=pyhd8ed1ab_0
jsonschema-specifications=2025.4.1=pyh29332c3_0
kiwisolver=1.4.2=py39hf939315_1
krb5=1.20.1=h143b758_1
lcms2=2.16=h92b89f2_1
ld_impl_linux-64=2.40=h12ee557_0
lerc=4.0.0=h6a678d5_0
libabseil=20250127.0=cxx17_h6a678d5_0
libblas=3.9.0=31_h59b9bed_openblas
libboost=1.82.0=h109eef0_2
libbrotlicommon=1.0.9=h166bdaf_7
libbrotlidec=1.0.9=h166bdaf_7
libbrotlienc=1.0.9=h166bdaf_7
libcblas=3.9.0=31_he106b2a_openblas
libclang13=14.0.6=default_he11475f_2
libcups=2.4.2=h2d74bed_1
libcurl=8.12.1=hc9e6f67_0
libdeflate=1.22=h5eee18b_0
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h516909a_1
libevent=2.1.12=hdbd6064_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=13.2.0=h69a702a_0
libgfortran5=13.2.0=ha4646dd_0
libglib=2.78.4=hdc74915_0
libgomp=11.2.0=h1234567_1
libgrpc=1.71.0=h2d74bed_0
libiconv=1.17=h166bdaf_0
liblapack=3.9.0=31_h7ac8fdf_openblas
liblapacke=3.9.0=31_he2f377e_openblas
libllvm14=14.0.6=hecde1de_4
libnghttp2=1.57.0=h2d74bed_0
libopenblas=0.3.29=ha39b09d_0
libpng=1.6.39=h5eee18b_0
libpq=17.4=hdbd6064_0
libprotobuf=5.29.3=h3cdef7c_1
libre2-11=2024.07.02=h6a678d5_0
libssh2=1.11.1=h251f7ec_0
libstdcxx-ng=11.2.0=h1234567_1
libthrift=0.15.0=h1795dd8_2
libtiff=4.7.0=hde9077f_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_1
libxcb=1.17.0=h9b100fa_0
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.13.8=hfdd30dd_0
libxslt=1.1.41=h097e994_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=2.1.1=py39hb9d737c_1
matplotlib=3.9.2=py39hf3d152e_2
matplotlib-base=3.9.2=py39hbfdbfaf_1
minizip=4.0.3=hf59b114_0
mpmath=1.3.0=pypi_0
munkres=1.1.4=pyhd8ed1ab_1
mysql=8.4.0=h721767e_2
narwhals=1.44.0=pyhe01879c_0
ncurses=6.4=h6a678d5_0
networkx=3.2.1=pypi_0
nomkl=3.0=0
nspr=4.35=h6a678d5_0
nss=3.89.1=h6a678d5_0
numexpr=2.10.1=py39hd28fd6d_0
numpy=2.0.1=py39heeff2f4_1
numpy-base=2.0.1=py39h8a23956_1
openblas=0.3.29=h06a4308_0
openblas-devel=0.3.29=ha39b09d_0
openjpeg=2.5.2=h0d4d230_1
openldap=2.6.4=h42fbc30_0
openssl=3.0.16=h5eee18b_0
orc=2.1.1=hd396ef6_0
packaging=25.0=pyh29332c3_1
pandas=2.2.3=py39h6a678d5_0
pcre2=10.42=hebb0a14_1
pillow=11.1.0=py39hac6e08b_1
pip=25.1=pyhc872135_2
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_2
plotly=6.1.2=pyhd8ed1ab_0
protobuf=5.29.3=py39he621ea3_0
pthread-stubs=0.3=h0ce48e5_1
pyarrow=19.0.0=py39h6a678d5_1
pycparser=2.22=pyh29332c3_1
pydeck=0.9.1=pyhd8ed1ab_0
pyparsing=3.2.3=pyhd8ed1ab_1
pyside6=6.7.2=py39ha2c6bb1_0
pysocks=1.7.1=pyha55dd90_7
python=3.9.23=he99959a_0
python-dateutil=2.9.0.post0=pyhff2d567_1
python-tzdata=2025.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytz=2025.2=pyhd8ed1ab_0
pyyaml=6.0=py39hb9d737c_4
qtbase=6.7.3=hdaa5aa8_0
qtdeclarative=6.7.3=h6a678d5_0
qtshadertools=6.7.3=h6a678d5_0
qtsvg=6.7.3=he621ea3_0
qttools=6.7.3=h80c7b02_0
qtwebchannel=6.7.3=h6a678d5_0
qtwebengine=6.7.3=h7abb84b_0
qtwebsockets=6.7.3=h6a678d5_0
re2=2024.07.02=hdb19cb5_0
readline=8.2=h5eee18b_0
referencing=0.36.2=pyh29332c3_0
regex=2024.11.6=pypi_0
requests=2.32.4=pyhd8ed1ab_0
rpds-py=0.22.3=py39h4aa5aa6_0
s2n=1.3.27=hdbd6064_0
safetensors=0.5.3=pypi_0
setuptools=78.1.1=py39h06a4308_0
six=1.17.0=pyhd8ed1ab_0
smmap=5.0.2=pyhd8ed1ab_0
snappy=1.2.1=h6a678d5_0
sqlite=3.45.3=h5eee18b_0
streamlit=1.46.0=pyhd8ed1ab_0
sympy=1.13.3=pypi_0
tenacity=9.1.2=pyhd8ed1ab_0
tk=8.6.14=h993c535_1
tokenizers=0.13.3=pypi_0
toml=0.10.2=pyhd8ed1ab_1
torch=2.0.1+cpu=pypi_0
torchvision=0.15.2+cpu=pypi_0
tornado=6.1=py39hb9d737c_3
tqdm=4.67.1=pypi_0
transformers=4.30.2=pypi_0
typing-extensions=4.14.0=h32cad80_0
typing_extensions=4.14.0=pyhe01879c_0
tzdata=2025b=h04d1e81_0
unicodedata2=14.0.0=py39hb9d737c_1
urllib3=2.5.0=pyhd8ed1ab_0
utf8proc=2.6.1=h5eee18b_1
watchdog=6.0.0=py39hf3d152e_0
wheel=0.45.1=py39h06a4308_0
xcb-util=0.4.1=h5eee18b_2
xcb-util-cursor=0.1.5=h5eee18b_0
xcb-util-image=0.4.0=h5eee18b_2
xcb-util-renderutil=0.3.10=h5eee18b_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xz=5.6.4=h5eee18b_1
yaml=0.2.5=h7f98852_2
zipp=3.23.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_1
zstandard=0.23.0=py39h2c38b39_1
zstd=1.5.6=hc292b87_0
